refactor: move RecordingAppender to zeebe-test-util
