+++ zeebe/test-util/pom.xml
@@ -196,6 +196,11 @@
       <artifactId>jmh-core</artifactId>
     </dependency>
 
+    <dependency>
+      <groupId>org.apache.logging.log4j</groupId>
+      <artifactId>log4j-core</artifactId>
+    </dependency>
+
   </dependencies>
   <build>
     <plugins>
