+++ zeebe/test-util/src/main/java/io/camunda/zeebe/test/util/logging/RecordingAppender.java
@@ -5,7 +5,7 @@
  * Licensed under the Camunda License 1.0. You may not use this file
  * except in compliance with the Camunda License 1.0.
  */
-package io.camunda.zeebe.util.logging;
+package io.camunda.zeebe.test.util.logging;
 
 import java.io.Serializable;
 import java.util.ArrayList;
